Rewriting attempt and ranking collection

+++ Backend/src/modules/live_data_processor/material/server.rs
@@ -20,7 +20,7 @@ pub struct Server {
     pub instance_participants: HashMap<u32, BTreeSet<u32>>,
     // Per instance there is a set of active attempts and when they began,
     // though most of the times only 1
-    pub active_attempts: HashMap<u32, HashMap<u64, Attempt>>,
+    pub active_attempts: HashMap<u32, HashMap<u32, Attempt>>,
 
     // Used to handle unordered events
     pub subject_prepend_mode_set: BTreeSet<u64>, // Contains server_uid of subject
